+++ CHANGES.rst
@@ -169,6 +169,9 @@ warning is emitted.
 
 - ``astropy.table``
 
+  - Fix a problem that caused MaskedColumn.__getitem__ to not preserve column
+    metadata. [#1471]
+
 - ``astropy.time``
 
   - Fix a problem when converting to one of the formats `unix`,